api/serverless_requirements.txt
requests==2.31.0
python-telegram-bot==20.8
python-dotenv==1.0.0 